Moved media file parser to its own module
