+++ core/build.gradle
@@ -25,6 +25,7 @@ dependencies {
     implementation project(':net:sync:gpoddernet')
     implementation project(':net:sync:model')
     implementation project(':parser:feed')
+    implementation project(':parser:media')
     implementation project(':ui:app-start-intent')
     implementation project(':ui:common')
     implementation project(':ui:png-icons')
