+++ core/src/main/java/de/danoeh/antennapod/core/storage/mapper/ChapterCursorMapper.java
@@ -3,10 +3,10 @@ package de.danoeh.antennapod.core.storage.mapper;
 import android.database.Cursor;
 import androidx.annotation.NonNull;
 import de.danoeh.antennapod.model.feed.Chapter;
-import de.danoeh.antennapod.core.feed.ID3Chapter;
 import de.danoeh.antennapod.parser.feed.element.SimpleChapter;
-import de.danoeh.antennapod.core.feed.VorbisCommentChapter;
 import de.danoeh.antennapod.core.storage.PodDBAdapter;
+import de.danoeh.antennapod.parser.media.id3.ID3Chapter;
+import de.danoeh.antennapod.parser.media.vorbis.VorbisCommentChapter;
 
 /**
  * Converts a {@link Cursor} to a {@link Chapter} object.
