+++ parser/media/README.md
@@ -1,0 +1,3 @@
+# :parser:media
+
+This module provides the tag parser for media files. This includes id3 or ogg/vorbis.
