+++ parser/media/src/main/java/de/danoeh/antennapod/parser/media/id3/ID3Reader.java
@@ -1,9 +1,9 @@
-package de.danoeh.antennapod.core.util.id3reader;
+package de.danoeh.antennapod.parser.media.id3;
 
 import android.util.Log;
 import androidx.annotation.NonNull;
-import de.danoeh.antennapod.core.util.id3reader.model.FrameHeader;
-import de.danoeh.antennapod.core.util.id3reader.model.TagHeader;
+import de.danoeh.antennapod.parser.media.id3.model.FrameHeader;
+import de.danoeh.antennapod.parser.media.id3.model.TagHeader;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.io.input.CountingInputStream;
