+++ parser/media/src/main/java/de/danoeh/antennapod/parser/media/id3/ID3ReaderException.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.core.util.id3reader;
+package de.danoeh.antennapod.parser.media.id3;
 
 public class ID3ReaderException extends Exception {
     private static final long serialVersionUID = 1L;
